fix(packaging,persistence): move CLI-needed repos under src/cp + ship in tarball
cli-v0.3.0's release tarball is unrunnable — `bun install -g` succeeds
but `ocpp-cp-sim --help` fails immediately with:

    Cannot find module '../data/sqlite/SqliteScenarioRepository'
        from '.../src/cli/service.ts'

src/cli/service.ts had grown imports for SqliteScenarioRepository (#51),
SqliteConnectorRuntimeRepository / NoopConnectorRuntimeRepository (#57)
and src/utils/scenarios/*.json (#54), but the `files` list in
package.json only ever shipped src/cli, src/cp,
src/utils/scenarioTemplates.ts, and dist. Docker copies every src/* path
(PR #52/#55), so the gap only bit users installing from the published
tarball.

The minimal fix would be to pin those paths into `files`. The deeper
issue is that the files were under `src/data/` to begin with — that
tree is otherwise React / sql.js / IndexedDB browser code, and pinning
specific paths inside it would just guarantee a repeat the next time
the CLI grew a dependency. Move the four files the daemon actually uses
out from under `src/data/` and under `src/cp/domain/persistence/`,
where the abstract Database and BunSqliteDatabase already live:

  src/data/interfaces/ScenarioRepository.ts             →
    src/cp/domain/persistence/ScenarioRepository.ts
  src/data/interfaces/ConnectorRuntimeRepository.ts     →
    src/cp/domain/persistence/ConnectorRuntimeRepository.ts
  src/data/sqlite/SqliteScenarioRepository.ts           →
    src/cp/domain/persistence/SqliteScenarioRepository.ts
  src/data/sqlite/SqliteConnectorRuntimeRepository.ts   →
    src/cp/domain/persistence/SqliteConnectorRuntimeRepository.ts

Their imports only ever named the abstract Database, OcppTypes, and the
Transaction domain object — all already in src/cp/. Browser consumers
(DataProvider, LocalChargePointService) and the BunSqliteDatabase test
were the only callers that needed import-path updates.

With the move in place, package.json `files` drops the per-file pinning
back to just `src/cli`, `src/cp`, `src/utils/scenarioTemplates.ts`,
`src/utils/scenarios`, `dist`. A future CLI import from `src/data` can
no longer accidentally re-introduce the v0.3.0-style install break.

Verified:
- vite build succeeds (browser bundle picks up the new paths).
- bun test on BunSqliteDatabase.bun.test.ts passes (8/8).
- npm pack --dry-run lists zero src/data/** files; the four moved files
  appear under src/cp/domain/persistence/.
- Daemon smoke: POST /v1/cp seeds Essential CP Behavior on both
  connectors (essential-cp-behavior-E1-c{1,2}-…).
