+++ src/cp/domain/persistence/ScenarioRepository.ts
@@ -1,8 +1,15 @@
-import type { ScenarioDefinition } from "../../cp/application/scenario/ScenarioTypes";
+import type { ScenarioDefinition } from "../../application/scenario/ScenarioTypes";
 
 export interface ScenarioRepository {
-  load(chargePointId: string, connectorId: number | null): Promise<ScenarioDefinition | null>;
-  save(chargePointId: string, connectorId: number | null, scenario: ScenarioDefinition): Promise<void>;
+  load(
+    chargePointId: string,
+    connectorId: number | null,
+  ): Promise<ScenarioDefinition | null>;
+  save(
+    chargePointId: string,
+    connectorId: number | null,
+    scenario: ScenarioDefinition,
+  ): Promise<void>;
   delete(chargePointId: string, connectorId: number | null): Promise<void>;
   list(chargePointId: string): Promise<ScenarioDefinition[]>;
   subscribe(
